使用 figaro 实现工程配置
