+++ .gitignore
@@ -29,3 +29,6 @@ public/assets/
 .\#*
 .DS_*
 ._*
+
+# Ignore application configuration
+/config/application.yml
